fix: Settings route protected
