+++ package.json
@@ -15,7 +15,7 @@
     "react-dom": "^18.0.0",
     "react-icons": "^4.9.0",
     "react-redux": "^8.1.2",
-    "react-router-dom": "^6.16.0",
+    "react-router-dom": "^6.20.0",
     "react-scripts": "^5.0.1",
     "react-spinners": "^0.13.8",
     "redux-thunk": "^2.4.2",
