+++ src/App.js
@@ -1,25 +1,6 @@
-import { RouterProvider, createBrowserRouter } from 'react-router-dom';
+import { RouterProvider } from 'react-router-dom';
 
-import RootLayout from './pages/Root';
-import HomePage from './pages/Home';
-import CardPage from './pages/Card';
-import SignInPage from './pages/SignIn';
-import ErrorPage from './pages/Error';
-import SettingsPage from './pages/Settings';
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <RootLayout />,
-    errorElement: <ErrorPage />,
-    children: [
-      { index: true, element: <HomePage /> },
-      { path: 'signin', element: <SignInPage /> },
-      { path: 'settings', element: <SettingsPage /> },
-      { path: 'card/:id', element: <CardPage /> },
-    ],
-  },
-]);
+import router from './routes/routes';
 
 function App() {
   return <RouterProvider router={router} />;
