+++ src/routes/protectedRoute.js
@@ -1,0 +1,14 @@
+import { Navigate } from 'react-router-dom';
+import { useSelector } from 'react-redux';
+
+const ProtectedRoute = ({ children }) => {
+  const isLoggedIn = useSelector((state) => state.login.isLoggedIn);
+  const isAdmin = useSelector((state) => state.login.isAdmin);
+
+  if (!isLoggedIn || !isAdmin) {
+    return <Navigate to="/" replace />;
+  }
+  return children;
+};
+
+export default ProtectedRoute;
